chore(lvgl-pc): upgrade LVGL library to latest version
- Update LVGL core headers with new private API separation
- Add private header files for internal implementation details (lv_*_private.h)
- Remove deprecated nanovg and nuttx mock headers
- Add new hardware acceleration support for PPE and RTK platforms
- Update draw system with new private headers for sw, dma2d, and eve backends
- Reorganize display and driver headers with private API isolation
- Update CMakeLists.txt and configuration files for new build structure
- Modernize API encapsulation by separating public and private interfaces

+++ lvgl-pc/lvgl-lib/include/lvgl/src/misc/lv_rb.h
@@ -25,31 +25,32 @@ extern "C" {
  *      TYPEDEFS
  **********************/
 
-typedef enum {
+typedef enum
+{
     LV_RB_COLOR_RED,
     LV_RB_COLOR_BLACK
 } lv_rb_color_t;
 
-typedef int32_t lv_rb_compare_res_t;
+typedef int8_t lv_rb_compare_res_t;
 
-typedef lv_rb_compare_res_t (*lv_rb_compare_t)(const void * a, const void * b);
+typedef lv_rb_compare_res_t (*lv_rb_compare_t)(const void *a, const void *b);
 
 /**********************
  * GLOBAL PROTOTYPES
  **********************/
 
-bool lv_rb_init(lv_rb_t * tree, lv_rb_compare_t compare, size_t node_size);
-lv_rb_node_t * lv_rb_insert(lv_rb_t * tree, void * key);
-lv_rb_node_t * lv_rb_find(lv_rb_t * tree, const void * key);
-void * lv_rb_remove_node(lv_rb_t * tree, lv_rb_node_t * node);
-void * lv_rb_remove(lv_rb_t * tree, const void * key);
-bool lv_rb_drop_node(lv_rb_t * tree, lv_rb_node_t * node);
-bool lv_rb_drop(lv_rb_t * tree, const void * key);
-lv_rb_node_t * lv_rb_minimum(lv_rb_t * node);
-lv_rb_node_t * lv_rb_maximum(lv_rb_t * node);
-lv_rb_node_t * lv_rb_minimum_from(lv_rb_node_t * node);
-lv_rb_node_t * lv_rb_maximum_from(lv_rb_node_t * node);
-void lv_rb_destroy(lv_rb_t * tree);
+bool lv_rb_init(lv_rb_t *tree, lv_rb_compare_t compare, size_t node_size);
+lv_rb_node_t *lv_rb_insert(lv_rb_t *tree, void *key);
+lv_rb_node_t *lv_rb_find(lv_rb_t *tree, const void *key);
+void *lv_rb_remove_node(lv_rb_t *tree, lv_rb_node_t *node);
+void *lv_rb_remove(lv_rb_t *tree, const void *key);
+bool lv_rb_drop_node(lv_rb_t *tree, lv_rb_node_t *node);
+bool lv_rb_drop(lv_rb_t *tree, const void *key);
+lv_rb_node_t *lv_rb_minimum(lv_rb_t *node);
+lv_rb_node_t *lv_rb_maximum(lv_rb_t *node);
+lv_rb_node_t *lv_rb_minimum_from(lv_rb_node_t *node);
+lv_rb_node_t *lv_rb_maximum_from(lv_rb_node_t *node);
+void lv_rb_destroy(lv_rb_t *tree);
 
 /*************************
  *    GLOBAL VARIABLES
